instagram/requirements.txt
anthropic==0.34.0
openai==1.51.0
httpx==0.27.2
requests==2.32.3
pillow==10.4.0
python-dotenv==1.0.1
schedule==1.2.2
tweepy==4.14.0
